Removed trigger-deployment.yml because log files are deleted every time a new build is triggered on external hosts.
trigger-deployment.yml was for committing a useless file every x minutes. Meant to be used when using external hosting sites that do not support cron jobs (like railway.app), but support running a new build and deploy it whenever there is a commit.

Not viable because log files, which is needed for checking if a URL has already been tweeted out, gets deleted every time a new build is triggered due to a commit. So using GitHub actions instead of integrating a persistent database like PostgreSQL or MySQL.