Added working YAML file for GitHub actions (ritter-bot.yml)
ritter-bot.yml runs the app.py file every hour using GitHub actions, then commits the log_files folder to a 'deployment' branch of the repo.

The next time it runs, it checks out the deployment branch, meaning log_files is preserved instead of getting deleted with every build.

Should add an if condition that checks if a branch called 'deployment' exists in the repository.