Changed the workflow to run on 2 jobs.
The first job checks if a deployment branch exists. If not, it creates it and pushes it to origin.
The second job runs the script and pushes the log_files to the deployment branch on origin.

Changed "\\" to "/" in file path to make it work on Ubuntu VMs. This was the reason logging wasn't working since "file_name\\directory" isn't a valid path on Linux.

Now logging tweeting to the console to see tweet IDs, tweeted URLs and basic error info on the console when running. Also changed tld() to tldize().

Changed default mode to tweeting out title+url only.